+++ nisystemlink/clients/assetmanagement/_asset_management_client.py
@@ -18,7 +18,7 @@ from . import models
     on_exception=retry.CONNECTION_ERROR,
 )
 class AssetManagementClient(BaseClient):
-    def __init__(self, configuration: Optional[HttpConfiguration] = None):
+    def __init__(self, configuration: HttpConfiguration | None = None):
         """Initialize an instance.
 
         Args:
@@ -127,7 +127,7 @@ class AssetManagementClient(BaseClient):
     @post("assets/{assetId}/file", args=[Path("assetId"), Field("fileIds")])
     def link_files(
         self, asset_id: str, file_ids: List[str]
-    ) -> Optional[models.LinkFilesPartialSuccessResponse]:
+    ) -> models.LinkFilesPartialSuccessResponse | None:
         """Link files to an asset.
 
         Args: