+++ nisystemlink/clients/dataframe/models/_append_table_data_request.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 from nisystemlink.clients.core._uplink._json_model import JsonModel
 
 from ._data_frame import DataFrame
@@ -10,8 +8,8 @@ class AppendTableDataRequest(JsonModel):
     required unless ``endOfData`` is true.
     """
 
-    frame: Optional[DataFrame] = None
+    frame: DataFrame | None = None
     """The data frame containing the rows to append."""
 
-    end_of_data: Optional[bool] = None
+    end_of_data: bool | None = None
     """Whether the table should expect any additional rows to be appended in future requests."""