+++ nisystemlink/clients/dataframe/models/_modify_tables_request.py
@@ -1,4 +1,4 @@
-from typing import Dict, List, Optional
+from typing import Dict, List
 
 from nisystemlink.clients.core._uplink._json_model import JsonModel
 
@@ -11,7 +11,7 @@ class TableMetadataModification(JsonModel):
     id: str
     """The ID of the table to modify."""
 
-    metadata_revision: Optional[int] = None
+    metadata_revision: int | None = None
     """When specified, this is an integer that must match the last known
     revision number of the table, incremented by one. If it doesn't match the
     current ``metadataRevision`` incremented by one at the time of execution, the
@@ -19,21 +19,21 @@ class TableMetadataModification(JsonModel):
     ensure that changes to this table's metadata are based on a known, previous
     state."""
 
-    name: Optional[str] = None
+    name: str | None = None
     """The new name of the table."""
 
-    test_result_id: Optional[str] = None
+    test_result_id: str | None = None
     """The new test result ID associated with the table. Set to an empty string to remove
     the test result ID. Added in version 2 of the
     :py:attr:`nisystemlink.clients.dataframe.models.OperationsV1.modify_metadata` operation. Older
     versions of the service will ignore this value."""
 
-    workspace: Optional[str] = None
+    workspace: str | None = None
     """The new workspace for the table. Changing the workspace requires
     permission to delete the table in its current workspace and permission to
     create the table in its new workspace."""
 
-    properties: Optional[Dict[str, Optional[str]]] = None
+    properties: Dict[str, str | None] | None = None
     """The properties to modify. A map of key value properties containing the
     metadata to be added or modified. Setting a property value to ``None`` will
     delete the property. Existing properties not included in the map are
@@ -46,5 +46,5 @@ class ModifyTablesRequest(JsonModel):
     tables: List[TableMetadataModification]
     """The table modifications to apply. Each table may only appear once in the list."""
 
-    replace: Optional[bool] = None
+    replace: bool | None = None
     """When true, existing properties are replaced instead of merged."""