+++ nisystemlink/clients/dataframe/models/_query_table_data_request.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import List
 
 from nisystemlink.clients.core._uplink._with_paging import WithPaging
 
@@ -9,12 +9,12 @@ from ._query_table_data_base import QueryTableDataBase
 class QueryTableDataRequest(QueryTableDataBase, WithPaging):
     """Contains the filtering and sorting options to use when querying table data."""
 
-    order_by: Optional[List[ColumnOrderBy]] = None
+    order_by: List[ColumnOrderBy] | None = None
     """A list of columns to order the results by. Multiple columns may be
     specified to order rows that have the same value for prior columns. The
     columns used for sorting do not need to be included in the columns list, in
     which case they are not returned. If ``order_by`` is not specified, then the
     order in which results are returned is undefined."""
 
-    take: Optional[int] = None
+    take: int | None = None
     """Limits the returned list to the specified number of results."""