+++ nisystemlink/clients/notebook/models/_notebook_metadata.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 from nisystemlink.clients.core._uplink._json_model import JsonModel
 
@@ -7,32 +7,32 @@ from nisystemlink.clients.core._uplink._json_model import JsonModel
 class NotebookMetadata(JsonModel):
     """Metadata for a notebook."""
 
-    id: Optional[str] = None
+    id: str | None = None
     """The ID of the notebook."""
 
-    name: Optional[str] = None
+    name: str | None = None
     """The name of the notebook."""
 
-    workspace: Optional[str] = None
+    workspace: str | None = None
     """The Id of the workspace containing the notebook."""
 
-    created_by: Optional[str] = None
+    created_by: str | None = None
     """The Id of the user that created the notebook."""
 
-    updated_by: Optional[str] = None
+    updated_by: str | None = None
     """The Id of the user that last updated the notebook."""
 
-    created_at: Optional[datetime] = None
+    created_at: datetime | None = None
     """The created timestamp (ISO8601 format)."""
 
-    updated_at: Optional[datetime] = None
+    updated_at: datetime | None = None
     """The last updated timestamp (ISO8601 format)."""
 
-    properties: Optional[Dict[str, str]] = None
+    properties: Dict[str, str] | None = None
     """A map of key value properties associated with the notebook."""
 
-    metadata: Optional[Dict[str, Any]] = None
+    metadata: Dict[str, Any] | None = None
     """The metadata of the notebook."""
 
-    parameters: Optional[Dict[str, Any]] = None
+    parameters: Dict[str, Any] | None = None
     """The parameters of the notebook."""