+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!--next-version-placeholder-->
 
+## v2.22.0 (2025-12-10)
+
+### Feature
+
+* Add optional take to DataFrame Service export table data request model ([#177](https://github.com/ni/nisystemlink-clients-python/issues/177)) ([`2932991`](https://github.com/ni/nisystemlink-clients-python/commit/293299130297ef1ef800ebd94c4d0ab97ae06977))
+
+## v2.21.0 (2025-12-10)
+
+### Feature
+
+* Rev minimum python version to 3.10 and add python 3.14 ([#168](https://github.com/ni/nisystemlink-clients-python/issues/168)) ([`d414904`](https://github.com/ni/nisystemlink-clients-python/commit/d4149043ac2f1eff97cb7cf6e65b771839bd6e3c))
+
 ## v2.20.0 (2025-12-01)
 
 ### Feature